Removes Debian Jessie backports repo
The Debian Stretch VM never should have included the jessie-backports
target, and its inclusion causes apt updates to fail, blocking
installation of any and all packages. Let's snip it out.

The "quiet: true" declaration is included for compatibility with the
Whonix templates, which are also based on Debian, but don't use
/etc/apt/sources.list. See the salt docs [0] for reference:

  quiet -- Do not raise any exceptions. E.g. ignore the fact that the
  file that is tried to be edited does not exist and nothing really
  happened.

[0] https://docs.saltstack.com/en/latest/ref/states/all/salt.states.file.html#salt.states.file.line