Updates docs for Qubes 4.0 -> 4.0.1
Removes README ad-hoc steps for fedora-29

As of Qubes 4.0.1, fedora-29 is automatically installed, so no manual
steps are required to configure the common `sys-*` VMs with fedora-29
as base.

Removes ad-hoc commands for whonix-14 templates

As of Qubes 4.0.1, the default base templates for the Whonix VMs are the
14 series, so the out of the box config is what we want, allowing us to
drop the custom commands. Furthermore, the old templates are not
installed on 4.0.1, so we don't need to remove those, either.