Ensures salt config rpm removed in dev env
If environment=dev, then we DON'T want the salt RPM installed in dom0,
because it'll conflict with the direct-copy-from-git-repo strategy used
by the "dev" environment. So if we're installing from dev, let's make
sure that package is absent.

We might want to do this earlier in the run, e.g. in the Makefile, but I
think running it at all is good enough.